+++ CHANGELOG.md
@@ -8,6 +8,12 @@
   only be considered configured when the configuring declaration is upstream of
   the `!default` declaration.
 
+* When namespacing a negated variable, adds parentheses around it to prevent the
+  `-` from being parsed as part of the namespace.
+
+* Fix a bug in the migrating of removed color functions when the amount is a
+  variable being namespaced.
+
 ## 1.0.0
 
 * Initial release.